Move tooltip to proper place
