+++ packages/app/.storybook/DevContainer.tsx
@@ -1,6 +1,6 @@
 import { queryClient } from '@/config/query-client'
 import { getConfig } from '@/config/wagmi'
-import { TooltipProvider } from '@/ui/atoms/new/tooltip/Tooltip'
+import { TooltipProvider } from '@/ui/atoms/tooltip/Tooltip'
 import { RainbowKitProvider, lightTheme } from '@rainbow-me/rainbowkit'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { Suspense } from 'react'
