+++ packages/app/src/features/actions/settings-dialog/components/SettingsDialog.tsx
@@ -2,7 +2,7 @@ import { MultiPanelDialog } from '@/features/dialogs/common/components/MultiPane
 import { Dialog, DialogContent, DialogTitle, DialogTrigger } from '@/ui/atoms/dialog/Dialog'
 import { IconButton } from '@/ui/atoms/new/icon-button/IconButton'
 import { Switch } from '@/ui/atoms/new/switch/Switch'
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/new/tooltip/Tooltip'
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/tooltip/Tooltip'
 import { testIds } from '@/ui/utils/testIds'
 import { PortalRef } from '@/ui/utils/usePortalRef'
 import { Settings } from 'lucide-react'
