+++ packages/app/src/features/compliance/components/TermsOfService.tsx
@@ -4,8 +4,8 @@ import { MultiPanelDialog } from '@/features/dialogs/common/components/MultiPane
 import { assets } from '@/ui/assets'
 import { Link } from '@/ui/atoms/link/Link'
 import { Button } from '@/ui/atoms/new/button/Button'
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/new/tooltip/Tooltip'
 import { ScrollArea } from '@/ui/atoms/scroll-area/ScrollArea'
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/tooltip/Tooltip'
 import { links } from '@/ui/constants/links'
 import { cn } from '@/ui/utils/style'
 import { useIsIntersecting } from '@/ui/utils/useIntersecting'
