+++ packages/app/src/features/markets/components/summary-tile/components/Tile.tsx
@@ -2,7 +2,7 @@ import { HelpCircle } from 'lucide-react'
 
 import { NormalizedUnitNumber } from '@/domain/types/NumericValues'
 import { USD_MOCK_TOKEN } from '@/domain/types/Token'
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/new/tooltip/Tooltip'
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/tooltip/Tooltip'
 import { cn } from '@/ui/utils/style'
 import { ComponentType } from 'react'
