+++ packages/app/src/ui/molecules/new/copy-button/CopyButton.tsx
@@ -1,4 +1,4 @@
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/new/tooltip/Tooltip'
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/tooltip/Tooltip'
 import { cn } from '@/ui/utils/style'
 import { useClipboard } from '@/utils/useClipboard'
 import { useMounted } from '@/utils/useMounted'
