+++ packages/app/src/ui/organisms/new/transaction-overview/TransactionOverview.tsx
@@ -1,5 +1,5 @@
 import { assets } from '@/ui/assets'
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/new/tooltip/Tooltip'
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/tooltip/Tooltip'
 import { cn } from '@/ui/utils/style'
 import { InfoIcon } from 'lucide-react'
 import { HTMLAttributes } from 'react'
